Moved calculateDues to helper and added dues to WFC export
