+++ WcaOnRails/app/controllers/competitions_controller.rb
@@ -411,32 +411,16 @@ class CompetitionsController < ApplicationController
   end
 
   def calculate_dues
-    country_iso2 = Country.find_by(id: params[:country_id])&.iso2
-    country_band = CountryBand.find_by(iso2: country_iso2)&.number
-
-    update_exchange_rates_if_needed
-    input_money_us_dollars = Money.new(params[:entry_fee_cents].to_i, params[:currency_code]).exchange_to("USD").amount
-
-    registration_fee_dues_us_dollars = input_money_us_dollars * CountryBand::PERCENT_REGISTRATION_FEE_USED_FOR_DUE_AMOUNT
-    country_band_dues_us_dollars = CountryBand::BANDS[country_band][:value] if country_band.present? && country_band > 0
+    competition = Competition.find(params[:competition_id])
+    dues_per_competitor_in_usd = ApplicationController.helpers.dues_per_competitor_in_usd(competition)
 
     # times 100 because later currency conversions require lowest currency subunit, which is cents for USD
-    price_per_competitor_us_cents = [registration_fee_dues_us_dollars, country_band_dues_us_dollars].compact.max * 100
-
-    if ActiveRecord::Type::Boolean.new.cast(params[:competitor_limit_enabled])
-      estimated_dues_us_cents = price_per_competitor_us_cents * params[:competitor_limit].to_i
-      estimated_dues = Money.new(estimated_dues_us_cents, "USD").exchange_to(params[:currency_code]).format
-
-      render json: {
-        dues_value: estimated_dues,
-      }
-    else
-      price_per_competitor = Money.new(price_per_competitor_us_cents, "USD").exchange_to(params[:currency_code]).format
-
-      render json: {
-        dues_value: price_per_competitor,
-      }
-    end
+    price_per_competitor_us_cents = dues_per_competitor_in_usd * 100
+    multiplier = competition.competitor_limit_enabled? ? competition.competitor_limit : 1
+    total_dues = price_per_competitor_us_cents * multiplier
+    render json: {
+      dues_value: Money.new(total_dues, "USD").exchange_to(competition.currency_code).format,
+    }
   end
 
   def show
@@ -636,12 +620,6 @@ class CompetitionsController < ApplicationController
 
   private def confirming?
     params[:commit] == "Confirm"
-  end
-
-  private def update_exchange_rates_if_needed
-    if !Money.default_bank.rates_updated_at || Money.default_bank.rates_updated_at < 1.day.ago
-      Money.default_bank.update_rates
-    end
   end
 
   private def competition_params
